Moved into proper package
