+++ application/src/main/java/org/togetherjava/tjbot/commands/bookmarks/BookmarksSystem.java
@@ -1,4 +1,4 @@
-package org.togetherjava.tjbot.commands.help;
+package org.togetherjava.tjbot.commands.bookmarks;
 
 import net.dv8tion.jda.api.EmbedBuilder;
 import net.dv8tion.jda.api.entities.MessageEmbed;
