+++ application/src/main/java/org/togetherjava/tjbot/commands/bookmarks/package-info.java
@@ -1,0 +1,12 @@
+/**
+ * This package offers all functionality for maintaining bookmarks. For example commands that let
+ * users bookmark a help thread, such as
+ * {@link org.togetherjava.tjbot.commands.bookmarks.BookmarksCommand}.
+ */
+@MethodsReturnNonnullByDefault
+@ParametersAreNonnullByDefault
+package org.togetherjava.tjbot.commands.bookmarks;
+
+import org.togetherjava.tjbot.annotations.MethodsReturnNonnullByDefault;
+
+import javax.annotation.ParametersAreNonnullByDefault;
